io.registry: decorate io identify for cache/list
the `gwpy.io.register_identifier` method is now decorated to automatically parse file-like and collections of file-like objects to pass a single file path to the underlying identifier function.

this should mean that each identifier doesn't have to parse the format of a file string, and also that all readers are expected to handle a cache or list of files, mainly via `gwpy.io.cache.file_list`.