port LIGO_LW to python-ligo-lw
this is a backwards-compatible update to use the new python-ligo-lw library for LIGO_LW I/O, but with fallbacks to `glue.ligolw` when reading old files.
the default write action is designed to still be the old format, but there is a warning there to announce that this will change in the future.