EventTable.read(): modify API for hdf5.snax format
This modifies the channel argument to be optional and allows
for more flexible behavior for reading SNAX files. In particular,
this allows reading multiple channels at once from a single hdf5
file by providing a list of channels. If channels are not specified,
this will attempt to read in all channels from the file.

In order to preserve the uniqueness across channels, a 'channel'
column is added into the EventTable during the read process.

Also update relevant unit tests for all modified behavior.