Address inconsistency in transfer_function between CSD and PSD averaging methods
This PR fixes incorrect results when no average method is supplied to the transfer_function method (I note here that attempting to supple the `method=` argument yields an error in the csd() method). Here we explicitly pass an averaging method to both the PSD and CSD methods so that the results are consistent.

Closes #1618

(cherry picked from commit f36eaaae0f600696d21ed52fed1b4573a6241a34)