Rebuild TimeSeries.get using a registry
This is a major rebuild of the get-me-data-anyway-you-can functionality, which utilises a registry of getters to construct an ordered list of candidate sources. Each one is tried and the first result is returned; if nothing works an ExceptionGroup is raised collecting all of the individual errors.

This required major rewrites to the gwdatafind, gwosc, and nds2 modules in gwpy.timeseries.io, but the whole thing is now much simpler and more coherent.